Major rework of MusicColor.py
This reorganized the code so it be less spaghetti like. Specifically replaced the while loops that crash the program to end it, it uses for loops that end it gracefully. more of the audio processing was move into the audio function, which I changed the name for clarity. Other variable names were also changed for clarity.

Not that I don't like spaghetti, I just don't like it in my code.